Namespace the code; split into singular files; add composer.json file; ignore ide/composer junk
