+++ composer.json
@@ -1,0 +1,13 @@
+{
+    "name": "coopernurse/barrister",
+    "description": "A JSON-RPC 2.0 client/server with contractual checks",
+    "license": "MIT",
+    "require": {
+        "php": ">=5.3.0"
+    },
+    "autoload": {
+        "psr-0": {
+            "coopernurse": "src/main"
+        }
+    }
+}
