+++ src/main/coopernurse/barrister/BInterface.php
@@ -1,0 +1,25 @@
+<?php
+
+namespace coopernurse\barrister;
+
+class BInterface {
+
+  /** @var BFunction[] */
+  public $functions;
+
+  function __construct($iface) {
+    $this->functions = array();
+    foreach ($iface->functions as $i=>$func) {
+      $this->functions[$func->name] = new BFunction($func);
+    }
+  }
+
+  /**
+   * @param string $name
+   * @return BFunction
+   */
+  function getFunction($name) {
+    return $this->functions[$name];
+  }
+
+}
