+++ src/main/coopernurse/barrister/Barrister.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace coopernurse\barrister;
+
+class Barrister {
+
+  function httpClient($url) {
+    return new Client(new HttpTransport($url));
+  }
+
+}
